% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_glmnet.R
\name{boot_combine}
\alias{boot_combine}
\title{Combine two boot_glmnet objects}
\usage{
boot_combine(fit1, fit2)
}
\arguments{
\item{fit1}{a boot_glmnet object}

\item{fit2}{a boot_glmnet object, which must have the same alpha, lambda, and number of rows as fit1, but may have any number of columns > 1}
}
\value{
a boot_glmnet object
}
\description{
Combine two boot_glmnet objects
}
