% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_glmnet.R
\name{boot_glmnet}
\alias{boot_glmnet}
\title{Fit bootstrap replicates using glmnet::glmnet}
\usage{
boot_glmnet(
  x,
  y,
  alpha,
  lambda,
  cluster = NULL,
  verbose = TRUE,
  R = 100,
  file = NULL,
  ...
)
}
\arguments{
\item{x}{a model matrix or sparse representation of the model matrix}

\item{y}{a vector of the dependent or outcome observations}

\item{alpha}{mixing parameter which ranges from 0 to 1, where 0 is ridge regression and 1 is LASSO}

\item{lambda}{penalty term}

\item{verbose}{prints current progress and expected time to completion if TRUE}

\item{R}{number of bootstrap replicates or bootstrap replicates per block if computing in parallel}
}
\value{
a boot_glmnet object
}
\description{
Fit bootstrap replicates using glmnet::glmnet
}
