% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_glmnet.R
\name{bootfit}
\alias{bootfit}
\title{Fit a bootstrap replicate}
\usage{
bootfit(x, y, alpha, lambda, t0, R, z, verbose, ...)
}
\arguments{
\item{x}{Model matrix or sparse model matrix}

\item{y}{Dependent or outcome variable}

\item{alpha}{mixing parameter ranging from 0 to 1, with 0 being ridge regression, 1 being the LASSO and values in-between being elastic net regression}

\item{lambda}{penalty term}

\item{t0}{starting time}

\item{R}{number of bootstrap replicates}

\item{z}{current z or R bootstrap replicate}

\item{verbose}{print replicate number, elapsed time, and expected time to complete if TRUE}
}
\value{
RMSE, r-squared, and a matrix of penalized coefficients
}
\description{
Fit a bootstrap replicate
}
