% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_glmnet.R
\name{coef.boot_glmnet_q}
\alias{coef.boot_glmnet_q}
\title{Return bootstrap coefficients and confidence intervals}
\usage{
\method{coef}{boot_glmnet_q}(obj, ...)
}
\arguments{
\item{obj}{a boot_glmnet_q object}
}
\value{
a data.frame of bootstrap coefficients
}
\description{
Return bootstrap coefficients and confidence intervals
}
