% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_glmnet.R
\name{mc_boot_glmnet}
\alias{mc_boot_glmnet}
\title{Fit n x R blocks of bootstrap replicates in parallel using future.apply}
\usage{
mc_boot_glmnet(n = 4L, x, y, alpha = 1, lambda, R = 100, verbose = TRUE, ...)
}
\arguments{
\item{n}{number of blocks to compute in parallel}

\item{x}{Model matrix or sparse model matrix}

\item{y}{Dependent or outcome variable}

\item{alpha}{mixing parameter ranging from 0 to 1, with 0 being ridge regression, 1 being the LASSO and values in-between being elastic net regression}

\item{lambda}{penalty term}

\item{R}{number of bootstrap replicates}

\item{verbose}{print replicate number, elapsed time, and expected time to complete if TRUE}

\item{t0}{starting time}

\item{z}{current z or R bootstrap replicate}
}
\value{
RMSE, r-squared, and a matrix of penalized coefficients
}
\description{
Fit n x R blocks of bootstrap replicates in parallel using future.apply
}
