% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_glmnet.R
\name{p.lasso}
\alias{p.lasso}
\title{Calculate proportion of bootstrap replicates where a predictor has been selected}
\usage{
p.lasso(obj, ...)
}
\arguments{
\item{obj}{a boot_glmnet or boot_glmnet_q object}
}
\value{
a data.frame with proportion of replicates where a predictor has been selected and an approximate confidence interval
}
\description{
Calculate proportion of bootstrap replicates where a predictor has been selected
}
