% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_glmnet.R
\name{quant.boot_glmnet}
\alias{quant.boot_glmnet}
\title{Compute bootstrap statistics}
\usage{
quant.boot_glmnet(obj, ci = 0.95)
}
\arguments{
\item{obj}{a boot_glmnet object}

\item{ci}{specified confidence interval between 0 and 1}
}
\value{
a boot_glmnet_q object
}
\description{
Compute bootstrap statistics
}
