% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{Simulate (uncorrelated) multivariate normal data of size N x p}
\usage{
sim(
  N = 100,
  clusters = 20,
  p = 10,
  ranef.sd = 0.1,
  eff.sd = 0.1,
  resid.sd = NULL,
  icc = 0.5,
  sparsity = 0.5,
  treat = NULL
)
}
\arguments{
\item{N}{number of observations (will adjust if N not a multiple of clusters)}

\item{clusters}{number of clusters}

\item{p}{size of N x p matrix corresponding to number of predictors}

\item{ranef.sd}{standard deviation of simulated cluster random intercepts, assuming normal distribution}

\item{eff.sd}{standard deviation of distribution of effect sizes (difference between treat = 1 and treat = 0), assuming normal distribution with mean = 0}

\item{resid.sd}{optional standard deviation of residuals (inferred from icc if not supplied)}

\item{icc}{optional intra-class correlation (inferred if resid.sd supplied)}

\item{sparsity}{1-proportion of non-zero coefficients (equivalently proportion of coefficients that are exactly zero)}

\item{treat}{vector of treatment group coefficients (defaults to sample(c(0,1), clusters, replace = TRUE))}

\item{z}{current z of R bootstrap replicate, provided by boot_glmnet()}
}
\value{
list with $data data frame of simulated data, $eff.size data frame of effect sizes and standardized effects sizes for p variables, $icc, $resid.sd
}
\description{
Simulate (uncorrelated) multivariate normal data of size N x p
}
